+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project is expected to follow [Semantic Versioning](https://semver.org/
 - Add support for functions returning a simple scalar value with a return statement.
 - Add support for procedure parameters, simple types and by value only for now.
 
+### [Changed]
+
+- Remove (`break` | `next`) finalizer from switch statement, now the statement sequence ends with `end`.
+
 ## [v0.0.9] - 2026-06-02
 
 ### [Added]